chatbot-allenamento/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
starlette>=0.27.0
openai>=1.3.0
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-readers-file>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-simple>=0.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
aiofiles>=23.1.0
tiktoken>=0.5.0
